+++ package.json
@@ -101,19 +101,19 @@
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^5.1.0",
     "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-prettier": "^3.4.1",
     "eslint-plugin-react": "^7.29.4",
     "eslint-plugin-react-hooks": "^4.3.0",
     "eslint-plugin-testing-library": "^3.6.0",
+    "husky": "^8.0.0",
+    "jest-environment-jsdom-sixteen": "^1.0.3",
     "msw": "^0.19.5",
+    "node-fetch": "^3.1.1",
     "node-sass": "^8.0.0",
     "prettier": "^1.19.1",
     "redux-mock-store": "^1.5.4",
-    "husky": "^8.0.0",
     "sinon": "^7.3.2",
-    "surge": "^0.21.6",
-    "jest-environment-jsdom-sixteen": "^1.0.3",
-    "node-fetch": "^3.1.1"
+    "surge": "^0.21.6"
   }
 }
